Add 2p/4p toggle in artifact comparison modal
